Support older GHC.
On Ubuntu Trusty 64bit with GHC 7.6.3 from the Ubuntu repos with large values for the number of slices, we get errors like this:

$ ./pi
Calculating PI with 5000000 slices
Stack space overflow: current size 8388608 bytes.
Use `+RTS -Ksize -RTS' to increase it.

This doesn't happen on:

Windows 7 32 bit with GHC 8.0.1 from Haskell-Platform
Ubuntu Xenial 64 bit with GHC 7.10.3-7 from Ubuntu Repos

So I'm going to assume that stack size behaviour has changed in later versions.  Indeed on the last platform I've observed the code the last platform exceed a memory footprint of many gigabytes, as to the ~8 megabyte stack size in the above error message.

This change consists of:

1) Adding -rtsopts to the Makefile.  This enables passing of run-time options to the ghc runtime system.

2) Options to the invokation of pi in run.sh that set the stack size to 1600000000 which is enough for the default number of slices.

This problem occurs with both the original messy version of the code, and my improved "more functional" version.