Add the mixed C#/Java version.
I've been of the opinion for a while that you can probably write a program
that with a stub library compiles in both C# compilers and Java compilers.

Well it turns out that there's an additional layer of abstraction you need
to do based around the fact that C# needs you to explicitly import the
System library before it knows about Strings, and Java need main() to be
passed a String array as an argument.

Therefore this implementation has:

1) Main body of code in code.mixed (sym-linked and compiled)

2) Wrappers in pi.java, pi.cs

3) Stub library in sys.java, sys.cs

All controlled by a makefile.