Fix method declarations for Clojure 1.8
So we do need :methods for Clojure 1.8. Worse, we need to make them static and
prepend them with - which breaks on Clojure 1.6.
So we need a slight cludge where there's a method without the - which calls the one with.

Make clean was also slightly broken if pi.jar wasn't up to date - fixed.